Move Include param to .rubocop.yml
Here we're moving the Include param that specifies the list of files
where a cop should run. The param was not working because it should
reside in the `.rubocop.yml` file and not in the rubocop todo file.

I have checked for other such instance and there are not any.