Fix dependency aggregation pagination when cursor contains nulls
The dependency aggregations currently do not correctly account for
nullability of cursor columns. If the cursor contains a null value, then
we will attempt to build a query that checks for
sbom_occurrences.highest_severity > NULL which always returns 0 rows as
NULL cannot be compared with > or <. This MR updates the pagination to
correctly consider nullability. Additionally, keyset_paginate is removed
from the outer query because the column conditions added by it are
redundant and cause the query to waste time performing additional scans.
Pagination headers are now manually added in the controller, because
they cannot be automatically added by the serializer without
keyset_paginate on the relation.