Add a mailroom CI job to validate config
This job only attempts to load the MailRoom config, but does not
actually try to start the process. This validation step would prevent
issues as https://gitlab.com/gitlab-org/gitlab/-/issues/468467.

Note that this test uses Bundler even though Omnibus doesn't use it. A
future improvement would be to run MailRoom with just the gem itself.