Do not wrap flex parent of verified badge and commit copy button
In the "Add or remove previously merged commits" modal, if the author of
the commit committed with a verified ssh key and the commit message is
long, then this can cause the verified badge and the commit copy button
to have an undesirable layout. This is easily fixed by making the flex
parent of the badge and the commit hash copy button not wrap.

Changelog: fixed