Authenticate agent requests using Gitlab-Agentk-Api-Request header
Currently requests from agentk use an `Authorization` header, which
contains the agent token. In some contexts this header is assumed to
contain an OAuth token, not an agent token, which can lead to errors
being misreported. To solve this we will migrate to a new header,
`Gitlab-Agentk-Api-Request`, that will be specific to agent tokens.

See https://gitlab.com/gitlab-org/gitlab/-/issues/406582.

Changelog: fixed