Add error message for failed association deletion
**Problem**

When destroying a project and a child association of the project fails
deletion, it raises an ActiveRecord::RecordNotDestroyed exception.
This exception is not rescued and as a result of this, a vague deletion
failure reason is stored for the project.

**Solution**

Rescue the ActiveRecord::RecordNotDestroyed exception raised when a
child association fails so a clearer reason for failure is returned

Changelog: added