E2E: Add logging of environment certs for mtls scenario
https://gitlab.com/gitlab-org/gitlab/-/issues/431474 notes that some
failures can occur in the MTLS scenario due to TLS error when gitaly
tries to access the /internal/allowed endpoints. In order to debug
this, let's first verify what certs are actually present on disk to
determine if the flakiness is arising from the test environment config
or from an application config perspective.