Fill out fields on InputType
CI inputs have more configuration options than are currently available
on the GraphQL InputType. This commit adds `description`, `regex`, and
`type` fields to make more of the configuration option values available.

It does not add an `options` field because the value for options is a
mix of types, which is not easy to represent in GraphQL.

Changelog: added
Issue: https://gitlab.com/gitlab-org/gitlab/-/issues/466669