Drop unnecessary bundler install in CI
This slightly speeds up the CI tasks since we no longer have to
download another version of bundler.

Starting with Bundler v2.3, Bundler automatically installs the
version needed in `Gemfile.lock` if it's not installed already:
https://bundler.io/blog/2022/01/23/bundler-v2-3.html.

We've already crossed the Bundler v2.3 mark: Ruby 3.1.4 ships with
Bundler v2.3.27, and Ruby 3.2.4 ships with Bundler v2.4.19.

Relates to
https://gitlab.com/gitlab-org/gitlab-build-images/-/merge_requests/819