New query for group dependency list page sorting by 'license'
This is the final sort being modified for this [epic][0]

This (sort-of) mirrors the work done in:

- [MR 154054][1] (added sorting based on `highest_severity`)
- [MR 154779][2] (added sorting based on `component_name`)
- [MR 154863][3] (added sorting based on `package_manager`)

Originally, the implementation plan of sorting by `license` called for
adding a new column to sort on.

In the MR to add the column, a reviewer [mentioned][4] we could use the
existing `jsonb` data for sorting.

This involves modifying the `AggregationsFinder` to be able to
generate SQL to use jsonb to access the first `spdx_identifier` in the
`licenses` column

[0]:https://gitlab.com/groups/gitlab-org/-/epics/12371
[1]:https://gitlab.com/gitlab-org/gitlab/-/merge_requests/154054
[2]:https://gitlab.com/gitlab-org/gitlab/-/merge_requests/154779
[3]:https://gitlab.com/gitlab-org/gitlab/-/merge_requests/154863
[4]:https://gitlab.com/gitlab-org/gitlab/-/merge_requests/155072#note_1942448739

---

MR: https://gitlab.com/gitlab-org/gitlab/-/merge_requests/155995
related to: https://gitlab.com/gitlab-org/gitlab/-/issues/450922
Changelog: added