Always generate Gitaly configuration variants
GitalySetup is generating two configuration files for Gitaly, one
with transactions enabled and one without them. This is done in CI
as the configuration files are setup in the `setup-test-env` job
before the actual rspec jobs run. The rspec jobs however contain
the environment variable configuring whether the job should run with
or without transactions. Both configuration files being pre-generated
then allows us to decide when starting Gitaly for a given job which
configuration file to use.

Currently this is done only in CI but we'll soon need both
configuration files locally as well to enable Workhorse tests to pick
the appropriate one to use. Generate both configuration variants
always and use the appropriate one when starting Gitaly based on the
`GITALY_TRANSACTIONS_ENABLED` environment variable. This also makes
it easier to switch between transactions enabled and disable locally.
Previously the configuration file had to be regenerated if the setting
was changed, and this didn't happen automatically based on the
environment variable.