Recreate records when moving issues
When moving an issue from one project to another, we create a new issue
in the other project. With it we also changed the references like
`epic_issues` or `work_item_parent_links` to the new issue.

The `work_item_parent_links` table uses `namespace_id` as sharding key.
This key also needs to be updated to the new project's namespace_id.
However, it's recommended to not change sharding keys but treat the
records as immutable and re-create them with the new references.

As we also want to add `namespace_id` as sharding key to `epic_issues`,
we now recreate the `epic_issue` and `work_item_parent_link` records
with the new issue reference. We already have a trigger on insert to set
the `namespace_id` to the one of the new project.

Changelog: fixed
EE: true