Remove Dangerfile for prettier
This Dangerfile dates back to when we gradually adopted prettier. Now
our code base is fully prettier compliant and we have a separate
static-analysis job for prettier _and_ we run it part of lefthook. So
this Danger job just adds noise to the already long danger output and
potentially takes a long time.

Let's remove it, it outlived it's purpose.