Simplify Web IDE show view
We currently have a FOSS/EE split within the IDE views. The only
difference is that the EE version also loads xterm.css. This is only
needed for the legacy web IDE.

We are moving the import of the file to the `use_new_web_ide?` block
inside the FOSS version, making the EE version redundant.

Another minor deviation is the import of `page_bundles/terminal.css`
instead. I don't think it's a big concern because it only contains
_this_ CSS rule:

```
#terminal > div {
  min-height: 450px;
}
```

See: https://gitlab.com/gitlab-org/gitlab/-/issues/469663