Skip approval updates for merged MRs
This callback is causing note imports to fail for projects with aproval
rules configured. When a merge request is imported, if
`approvals_before_merge` is not nil and the MR has already been merged,
the approval rule generated by `any_approver_rule` will fail validation,
and the subsequent call to `update_column` will result in an exception.

https://gitlab.com/gitlab-org/gitlab/-/merge_requests/138154
validates against updating approval rules for merged MRs, so in this
case it should be safe to skip the callback for merged MRs.

Changelog: fixed
EE: true