Add table to record placeholder contributions
During an import that uses the improved user mapping from
https://gitlab.com/groups/gitlab-org/-/epics/12378, initially, the users
associated with the imported data are "placeholder" users (a kind of
internal user) rather than the real users.

The placeholder user is recorded in
`Import::SourceUser#placeholder_user`

Later the `Import::SourceUser` is assigned a real user, the
`Import::SourceUser#assigned_to_user`,  after the real user has
authorized this. At this point, we need to change all imported data from
being associated with the `placeholder_user` to the (real)
`assigned_to_user`.

The table in this commit records what imported data should be updated
in this process.

https://gitlab.com/gitlab-org/gitlab/-/issues/443554

Changelog: added