Add note unification for epic and epic work item
This allows transition from epic to epic work item notes to
be seamless as we UNION notes from epic and epic work item.

This will allow us to switch from loading epic notes to loading
epic work item notes and deprecate legacy epic model.

This change also touches some of the system synthetic notes
that are related to label and state changes and are based on
resource_label_events and resource_state_events.