Separate lock popovers between DOM parsing and UI
The LockPopovers Vue component was made as a catch all way
to avoid manually passing DOM nodes to to attach popovers
on all locked elements in settings due to cascading
privilidges. This worked mostly fine in HAML but is
not a great pattern for Vue applications because we
are effectively breaking the components based architecture
of limiting side effects to the component itself.

This commit splits the DOM parsing utils (so that
it can still be used in HAML) from the presentational
vue component so that we can use only the UI component
in Vue applications by passing a list of $refs instead.

There should be no visual changes.