Retain resource access token bot users after revocation
This change is behind a new feature flag,
`retain_resource_access_token_user_after_revoke`.

Resource Access Tokens (a.k.a. Group and Project Access Tokens) are
implemented by creating "bot" users, giving it membership to the
resource, and generating a Personal Access Token for that bot. When a
Resource Access Token is revoked, the bot user is scheduled for
deletion.

When the bot user is deleted the revoked token no longer has any
relation to the resource. This makes it impossible to know what a
revoked token had access to in the past.

With this change, when the FF is enabled, the bot user will be retained
indefinitely.