Update confirm danger phrase
The styling for the confirmation phrase in
the confirm danger modal is updated. The
phrase is divided in two separate lines. The first
line is bold. Since it is a label for the form
input the tag has been changed from <p> to
<label>. The phrase itself is on a separate
line.
The phrase has been updated in the constants
file, and been extracted to the translation file.
The test case has been updated to test for
the new phrasing.