Remove usage of let_it_be in feature spec
Based on the failures on this specific feature spec, code that is
not meant to be run inside a transaction are getting run within
a transaction even if it's not being called inside a transaction
out of the spec.

Theory is that since using `let_it_be` will wrap the example
group within a transaction, that causes the flakiness.