Run as-if-foss pipeline before approval
Since the as-if-foss pipeline should only run what EE would run,
and we only run predictive tests in EE, we should also only run
predictive in as-if-foss. This should not be expensive to run.

This will also run full as-if-foss for those which EE run full first,
but maybe that's really what we want anyway?