Fix incorrect retry warning when merge trains are enabled
The widget incorrectly used the merge strategy instead of the actual
pipeline when setting the merge train status for the warning.

Although the code changes are in GitLab FOSS, the problem and fix are
EE-specific because merge trains are EE-only.

See https://gitlab.com/gitlab-org/gitlab/-/issues/348912

Changelog: fixed
EE: true