Adds migration deleting AbuseReport records from notes
AbuseReport records by design lack a value for their namespace_id field.
This makes it impossible to shard the `notes` table by namespace.
However, AbuseReports were only inserted into this table via code behind
a feature flag, which was never default enabled, and there are no
existing records of this type in the .com databases. Manage::Auth team
is confident that this data can be deleted, as they move ahead with a
parallel effort to create a new `abuse_reports_notes` table to improve
the data structure.