Skip sbom_occurrences update on pipeline id change
Do not update records of the sbom_occurrences
table when only the pipeline id or the commit sha
have changed. This reduces the frequency
of tuple updates for the sbom_occurrences table.

As a result of this performance improvement,
the Dependency List and API link to
the commit where a SBOM component
was first detected instead of the last one.

See https://gitlab.com/gitlab-org/gitlab/-/issues/442486