List selector component for groups can fetch project only groups
Add option to independently select if a fetched groups are scoped
to a project and if a dropdown to switch between scoped and
unscoped groups should be shown. This change allows to limit the
fetched groups and not allow the end user to change that scoping
at the same time.

Changelog: changed
EE: true