Fix meta.root_namespace not logging for groups API requests
Previously `meta.root_namespace` did not log for any API call for
`/api/v4/groups/:id` requests. This happened because of two issues:

1. `lib/api/api.rb` attempts to push the `namespace` in
`ApplicationContext` if `@group` is set, but this instance variable
was never set for those API calls. Instead of using `user_group`, add
this to `find_group!` so both CE and EE API calls will set the
context.

2. If the user is an admin, `bypass_session!` is called before
`@group` is set. This method logs a message calling
`ApplicationContext#to_lazy_hash`, which memoizes `namespace` to be
`nil` since `@group` was `nil`. The fix in the first issue also takes
care of this.

Changelog: fixed