Add a `pending_delete` column to mark namespaces deleted by user
The `Groups::DestroyService` takes a couple of seconds to delete the
group but this duration can vary significantly when the group is large.
Which sometimes results in transient errors when fetching a list of
groups containing the deleted group.

Here we're adding a new column called `pending_delete` that will store
the info that the group has been deleted by the user. We also have a
`pending_delete` column in the projects table for the same reason.

Changelog: other