Allow filtering Resource Access Tokens by state in the REST API
The REST API has and endpoint to fetch Group and Project Access Tokens.
The API, in its current state, returns all APIs associated with the
resource.

This MR introduces the ability to filter the response by a token's
state. The default is unchanged - all tokens associated with the
resource are returned. `inactive` includes expired and revoked tokens.

Note: Prior to https://gitlab.com/gitlab-org/gitlab/-/issues/462217+,
revoking a token via the API or UI called
`ResourceAccessToken::RevokeService`, which deleted the bot user and
therefore the association between the token and the resource. Revoking
the user directly in the Rails console did not remove the bot user. So
the API would return: `active + expired + revoked_via_the_console`. When
that feature is completed and the FF is enabled the API will return
`active + expired + revoked_via_any_means`. So while there is
technically a change to the API response, it is becoming more
consistent.

Changelog: changed