Move merge request files now shown warning to diff panel
We move the warning in merge request that tell
users some files are not shown to the diff panel
instead of the main panel. This in effect means that
when the files panel is opened, the alert is not above
both the files and the diff, but instead under the diff
side and so take only half of the space. This change is so
that the message is more contextualized in the user reading
experience of the code.

Changelog: changed