Make status and confidentiality text sr-only instead of hidden
The status and confidentiality text is hidden in issuables and
work items in small screens. For visual users, this is okay
as they can understand status and confidentiality from their icons.
However, these icons are hidden from screenreaders so we should
continue communicating the text to screenreaders in small screens.

Also update the dropdown placement attributes to non-deprecated
values.