Fix dependency aggregation pagination when sorting by license
Keyset pagination uses `relation[attribute_name]` to get the value to
insert into the cursor. When sorting by licenses, we want this to be the
`spdx_identifier` but currently it is getting the entire `licenses`
jsonb object. This causes a 500 error when trying to navigate to the
second page while sorting by license. To fix this, we make the
`spdx_identifier` available through the select statement as
`primary_license_spdx_identifier` and we change the name of the
`ColumnOrderDefinition` for licenses to this so that it will be
inserted into the cursor.