Update code owner syntax docs to fix incorrect example
The example for Code Owners Unparsable Sections is incorrect. It gives
`[Section name]{2} @group` in the second example, this is a valid
section header.

The first example does not include the closing bracket which makes the
section header invalid.

This change removes the `]{2} @group` and also introduces 2 new
subheadings for readability.

Closes https://gitlab.com/gitlab-org/gitlab/-/issues/470935